fails hex, hexup and ptr

+++ 42cursus/ft_printf/ft_printf_hexup.c
@@ -6,7 +6,7 @@
 /*   By: rgallego <rgallego@student.42madrid>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/22 19:46:06 by rgallego          #+#    #+#             */
-/*   Updated: 2021/10/22 20:24:04 by rgallego         ###   ########.fr       */
+/*   Updated: 2021/10/25 20:58:55 by rgallego         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,13 +14,13 @@
 
 int	ft_printf_hexup(int nb)
 {
-	unsigned long long	aux;
+	long long	aux;
 	int					n_char;
 
 	if (nb < 0)
 	{
 		n_char = write(1, "-", 1);
-		aux = nb * -1;
+		aux = (long long)nb * -1;
 	}
 	else
 	{
