+++ 42cursus/ft_printf/ft_putnbr_base.c
@@ -6,15 +6,15 @@
 /*   By: rgallego <rgallego@student.42madrid>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/22 19:15:13 by rgallego          #+#    #+#             */
-/*   Updated: 2021/10/22 20:23:54 by rgallego         ###   ########.fr       */
+/*   Updated: 2021/10/25 20:56:20 by rgallego         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libftprintf.h"
 
-int	ft_putnbr_base(unsigned long long nb, char *base, int len_base)
+int	ft_putnbr_base(long long nb, char *base, int len_base)
 {
-	if (nb < len_base)
+	if (nb < (long long)len_base)
 		return (write(1, &base[nb], 1));
 	else
 		return (ft_putnbr_base(nb / 10, base, len_base)
